Copy old schema as "first migration"
The existing RAMS app's database was based on a structure.sql file. We took
this structure, brought it into MySQL, made sure the migrations were up to
date, and then allowed Rails to parse the DB into a schema.rb whose contents
we can use as "Migration #1" in this app. From here, we can make modifications
via migrations.

Also, in case we need it, the structure of the original DB is preserved as
this migration.