Configure CircleCI
In order to run on CircleCI, we need to make sure we're set up correctly.
First, we use MySQL and not postgres. Second, we can't specify that we connct
to `localhost` because MySQL assumes that means we want to connect locally
through a socket (even though that's not how hostnames work), so we need to
specify `127.0.0.1`. Third, there is a critical update in `actionpack` so we
needed to upgrade Rails from 6.1.4 to 6.1.4.1.