Rename `People` table to `users`
The name `users` is in keeping with Rails' conventions of using lowercase
names that are pluralized, and using "User" as the modelis in keeping with the
convention of referring to the people who will be using the application as
"Users".

This migration also changes all of the `PeopleID` columns in various tables to
`user_id`, which fits the Rails convention of using the singular, lowercased
name of the model in question plus `_id`. And as part of the renaming, the
indices on each table that refer to `PeopleID` were automatically changed to
`user_id`.