Add component for the account creation form
This commit introduces the first `ViewComponent` in the application. The `Unauthenticated::CreateAccountFormComponent` accepts a user and is initialized with a form object (currently using dummy data, but will be real eventually). This component contains helper logic and maps to a file that displays a form to create a new account. It will be included in a page component in a future commit.