Install Foreman, migrate on deploy
heroku uses the Procfile format to specify how services are run, which is
managed in Ruby by the `foreman` gem. There is a special job type called
`release` which is run after the code is uploaded but before it's deployed.
This makes it the perfect time to run DB migrations, which will happen on
deploy and does not require manual intervention.