Allow the user to sign out
There was no way for the user to sign out once they've signed in. This commit
changes the `sign_out` route from Devise to be able to be accessed via the
DELETE and GET HTTP methods. It also adds a temporary link to the application
layout so the user can be signed out. And finally, it adds a test to confirm
that it works.