Add accessibility coverage to system specs
This commit adds test coverage for evaluating accessibility in the system specs via `axe-core-rspec`/`axe-core-capybara` and updates the existing authentication-related system tests to check for accessibility violations.

There were several violations that were raised and addressed:
- Missing lang attribute to the html element in application layout. [More info](https://dequeuniversity.com/rules/axe/4.2/html-has-lang?application=axeAPI)
- Missing landmarks/region (i.e. a main element). [More info](https://dequeuniversity.com/rules/axe/4.2/landmark-one-main?app
lication=axeAPI)
- Level-one heading violations in various Devise generated pages [More info](https://dequeuniversity.com/rules/axe/4.2/page-has-heading-one?application=axeAPI)

It also addresses issues that were cropping up in the system specs as a result of trying to match against and navigate to an absolute url from within the `AuthenticationFlow` page object.