Subclass Devise::PasswordsController for views
Devise comes with a lot of built-in and generated code. This code is mostly
modifiable without doing a lot of digging, but we're in a situation where we
want an "unauthenticated" section that has its own layout. This commit allows
us to subclass the Devise::PasswordsController so that we can use this layout
and so that we can modify the views with ViewComponents. There is no
functional difference caused by this commit, and the specs pass unaltered.