Display user registration form
This commit eliminates the concept and use of `components` in this context. After more examination and experimentation, we decided that the `ViewComponent` library does not fully meet our needs/would likely become challenging to work with in the way that we had planned.

It also adds the registration form markup to the Devise registrations/new view and moves the translations into the `devise.en.yml` file so they can be properly namespaced.