Styled CSS for Forgot Password page
Since the Unauthenticated views were unstyled this page contains the styled
layout for those views as well as a sample view, which is the Forgot Password
view.

This commit also changes the Forgot Password view into a normal view instead
of a Component. The Components are for resuable, dynamic, testable classes
that represent view functionality. This is not really what that view needed,
and so became a little unnecessarily complicated. So we backed it out and it's
simpler now.