Update countries table and add model
This commit renames the `countries` table and its columns in order to bring them to parity with rails naming conventions. It makes several other modifications to the database as well including adding a null constraint to an id that corresponds to a country in a different table, renaming some indicies, and renaming columns in other tables that reference what was previously `CountryID`.

It also adds the model, with a subset of validations and associations (the rest of which will be added soon as other models are built) and tests.