Update institutions table and add model
This commit renames the `institutions` table and its columns in order to bring them to parity with rails naming conventions. It makes several other modifications to the database as well including changing column names, adding null constraints to ids across several tables, renaming some indicies, and renaming columns in other tables that reference what was previously `InstitutionID`.

It also adds the model, with a subset of validations and associations (the rest of which will be added soon as other models are built) and tests and updates the `Country` and `State` models' associations and validations to include `institutions`.