Update users table and model
This commit renames the `users` table and its columns in order to bring them to parity with rails naming conventions. It makes several other modifications to the database as well including adding null constraints to ids across several tables, renaming indicies, and adding some new columns.

It also updates the model, to include validations and associations and tests. It also updates the `Institution` model's associations and validations to include `users`.