Implement styles for other Unauthenticated views
This commit makes sure the Sign In and Reset Password pages are also styled,
same as the Forgot Password page. The main effort was to make sure they route
to controllers that user the `unauthenticted` layout and, then, secondarily,
to make sure the markup was sufficient to look like the design (basically,
mirror the Forgot Password markup).