Implement the Welcome Modal on the Homepage
This is a slightly dummy implementation, as it shows every time the homepage
renders. This will not always be the case.

This commit renders the Welcome modal on the homepage, welcoming a new user to
RAMS. It also has a mobile-appropriate layout. This commit also fixes the
homepage to have a mobile-appropriate layout so the mobile Modal doesn't have
weird scrollbars (which it still does slightly, but isn't a huge deal at the
moment).