Add Accesssibility Concenrs to the Welcome Modal
Modals can still interact well with screenreaders. Following this guide:
http://web-accessibility.carnegiemuseums.org/code/dialogs/, this commit
attempts to make the modals more accessible. The aria-hidden attribute is
employed to make sure the text of the page is not going to be read when it's
not relevant, and also to make sure that the modal text is not read when it's
not visible. Also, it adds a non-visible description as to what the modal is
for.