Incorporate the Modal in the system specs
The Authentication system spec needs to have the modal dismissed in order to
function. The Regitration system spec needs to have the modal not pop up
unexpectedly in order to function.

This commit changes the JS ModalController from always opening the modal if
there is a need for the Controller to be initialized to always initializing
but only opening a modal if one is defined as a `dialog` target. This allows
the space to exist and only be used if needed.