Enabled Devise's Paranoid Mode
This mode make Devise act as though there is an email address in its system,
even if there is not. This means it will not leak information on whether a
particular email address exists in its database or not. While not a giant
information leak, it's good security hygiene to prevent such accidental
disclosures.

Also, the message for resetting the password ended up being long once it was
using the paranoid variant of the message, so both the wording was tweaked and
the CSS was tweaked to better position it.