CSS Fixes for Unauthenticated and Homepage ieeues
The Unauthenticated and Homepage designs required a little tweaking for large
and small page widths, as well as (in the Homepage's case) to get the content
aligned properly.

This commit also adds some exceptions to the `axe_clean` matcher because the
contrast checking was unable to notice that the text in question was not
display on a white background. It was failing because the contrast ratio was
too low (~1.4), but had no problems with the white text in the same area
(which would have been a 1:1 ratio). So until the inconsistency is addressed
either by the authors of axe-core-rspec or they inform us of how to correct
it, it needs to remain excepted (yes, this is disappointing).