Enable Devise's Paranoid Mode
This mode makes it so that queries that may leak information do not leak such
information. For example, when resetting a password, if the email address does
not exist in the system, Devise says "If this email address is in our system,
you'll get an email about it" instead of "This email address is not in our
system", which prevents someone external from gaining information about
whether an email address (which represents a username to pair with potentially
correct passwords, or information about a person is particular) exists in the
system.