Allow proper clicking near notifications and nav
Then notifications and navigation on the unauthenticated pages sometimes
overlap the main content, making it impossible to click on elements of the
forms. Notably, when the user signs out, it becomes impossible to click on the
`Email Address` field to sign in again, until the notification is gone.

The problem was an excess of bottom-padding, which we remove here.

Side note: The height of the main element was changed from `100vh` to `100%`
because `100vh` was only making the element as tall as the window. This led to
underflow problems when the window was smaller.