Refactor State Option Select Select JS Controller
Prior to this, the StateOptionSelectController would make an ajax request to
load State data in JSON format and the populate the State dropdown with those
options. The "Stimulus Way" to do this is to allow the server to render some
HTML that will contain the options that is simply inserted into the `select`.
It makes for a much less complex javascript module.

It also allows for a more generic component, as this will operate on more than
just the State Select dropdown. It will, for example, be used to populate the
list of Amenities available when scheduling a visit to a Reserve.