Value Projection as a Stimulus Controller
This commit will allow an input to "project" its value onto an ancestor
element. While this seems of low value, it means that we can effect changes in
the layout almost purely through CSS, which makes things simpler and faster.

While we can specify CSS for a child element, we cannot specify CSS based on
the value of a cousin (meaning, a sibling that isn't a direct sibling, but is
the descendant of a common parent). For example:

```html
<section>
  <div>
    <select id="select">
      <option value="Personal"/>
      <option value="Business"/>
    </select>
  </div>
  <div id="only-personal">...</div>
  <div id="only-business">...</div>
</section>
```

In the above situation, we cannot set the visibility of the `only-business`
and `only-personal` elements based on the value of the `select`. But this
controller would "project" the select's value onto the parent `section`, which
lets us define CSS rules for expressly this purpose. And you can see that this
is what we do in the content of this commit.