Test Stimulus Controllers with Jest
This commit installs the testing harness necessary to be able to test Stimulus
Controllers without spinning up a whole browser environment. It also adds
tests that demonstrate the ValueProjectionController.

The changes in this commit are modified slightly from the ones found here:
https://shime.sh/testing-stimulus

It also adds `yarn jest` to the default rake tast, so it will run both by
default ans in CI.