Convert InstitutionSearch to FollowupContent
Since the introduction of the FollowupContentController, we don't need the
whole InstitutionSearchController anymore. This commit refactors the typeahead
on the Institution field of the signup and also introduces a small controller
to copy the content of an element into an input (which is one of the things
the InstitutionSearch functionality did).