Install simple_calendar gem
This commit adds the [`simple_calendar`](https://github.com/excid3/simple_calendar) gem to the project per the documentation. This gem will allow us to display a calendar that will be used to create a custom date picker but will also eventually serve a variety of purposes in the app.