Add Amenities Controller, showing per-Reservation
This commit allows the Reservation dropdown on the Schedule a Visit page to
change the list of Amenities on the form. It also does an as-necessary
implementation of some models that it needs.

Also, it changes the institutions_index_presenter_spec to match how the same
thing is being done in the amenities_index_presenter_spec.