Amenities when Requesting a Visit
The commit adds the form part of the Amenities on the Request a Visit page. We
needed to add a combobox in order to present the rate categories for the
Amenity, so we overlaid a select with a text box. This is not fully styled, of
course, but it functions as it is supposed to.