Move JS Controller Tests
Instead of keeping all the JS tests in the `app/javascripts/controllers`
directory where here are clearly not controllers, this commit moves them into
`app/javascripts/__tests__` so it can contain a bit more structure like the
`spec` directory. This commit was made to maintain the connection of these
files because they're going to be refactored a bit.