Extract some of the JS setup
When writing a test for the CounterController Stimulus module, there arose a
problem when running all the specs in that file. Namely, there seemed to be
two instances of the controller, so `up` ran twice and made the value of the
output 2 instead of 1, this failing the test. After a bit of digging and being
pointed towards resources, the link
https://damonbauer.dev/posts/testing-stimulus-with-jest/ described some useful
helper functions to use. This commit implements those functions for the
existing 4 test files.