Select visit arrival and departure dates
This commit allows a user to select arrival and departure dates by clicking dates on a calendar that is displayed in a modal. The selected dates will be formatted and populate the corresponding input field. Dates in the past cannot be selected and if they are clicked, the corresponding input field will be set to an empty string. The 'disabled' state of past dates will become more apparent when styling is added.