Reserve-Specific Inputs
On the Request a Visit form, when the user selects a Reserve, there is some
Reserve-specific information that needs to be display and collected. The Alert
is displayed, if there is one. And the `Special Needs` are collected (as well
we a "special needs" message displayed, if there is one).