Use native HTML date and time pickers
This commit opts to use the native date and time pickers that come as a part of HTML5 instead of custom date and time pickers. This is because building something custom has eaten up a significant amount fo development time but does not represent a major feature. If there is time at the end of the engagement, a custom date picker will be revisited so the UI matches the branding but for now, this is a simple and robust solution that does not rely on Javascript. There are some concerns about the experience of IE users, as the `date` and `time` field `type`s are not supported but it gracefully degrades to text fields, which in combination with server-side validations, can handle capturing valid dates and times.