Make the first Visits form page more submittable
This commit doesn't make the page fully submittable and useful from a form
perspective. But in order for it to work, this needed to be done. This changes
some things like the arrival/departure dates are actually named start/end and
the rendered page has the `.visits-create` class instead of `.visits-new`.

This is not complete, but adds necessary changes.