Filter projects by status category
This commit allows users to filter by status category (not the enum values directly). When a user selects a status to filter by in the drop down, we perform a lookup on a hash constant in the `Project` model that maps the `status_category` to the actual `status` value. This uses the `followup_content_controller` to make a `GET` request with query params (the selected status category) to the `ProjectStatusesController#index` endpoint. The `status` from the query params is extracted and passed to the `ProjectStatusesIndexPresenter`, which handles determining which `projects` should be displayed based on the supplied `status_category`. Once the query is complete, the `projects` are mapped over and a `ProjectPresenter` is initialized for each `project` and that collection is passed to the view that renders a `collection` of `projects`. Some changes were made to the `ProjectPresenter` to determine which `date` value to display and whether or not to display a `reserve_name` based on the selected `status_category`.