Get the Request A Visit form submitting (page 1)
This is not yet a complete form, but it does make Page 1 of this form into a
"working" form. When the form is submitted, it will redisplay all of its
contents as though there were errors. This shows that all the inputs are sent
and processed correctly and can also be redisplayed. This required some
rearchitecting of the components of this form, including the display of the
Amenities.

All of the Visit-related presentation classes were placed under the `visit`
namespace, so they can be sequestered away from places where we may need
similarly-named objects for other forms.