Add `status`, refactor enums in Visit
The enums in the Visit model are not specified in a wholly satisfying way.
This commit changes them so that instead of being of the form `"string" =>
"the same string"`, it uses `symbol: "string"`, which is more Rails-y and
works better with ActiveRecord's creation of enum methods.

It also adds an enum for the status ("Temp", "Approved", etc).

The enums can be set via the "key" or the "value". That is, both of these work:

`visit.status = "Pending approval"`
`visit.status = :pending_approval`