Convert Amenity partial to use fields_for
The Amenity form partial for the Visit form was being built very manually.
because we need to display errors consistently, we should build it using a
form builder instead of trying to replicate that functionality by hand. This
commit lets the partial use a "form" generated via `fields_for` which both
standardizes and simplifies te markup.

The only gotcha is the check box, since the `check_box` helper wil generate a
check box and a hidden field to simulate the value of an "unchecked" box. so
in order to get the name right, we need to generate the name based on the
object_name of the form we're inside. But since this is using standard Rails
form builder methods, it shouldn't be a problem; it's just not the usual way.