Can submit Visit Page 1, see Visit-relevant errors
When you load the Request a Visit form and immediately press Submit, you will
see (almost) all errors relevant to the Visit object itself. The exceptions
are the error on the Project (since by default the project select is not
visible) and the error for not selecting a public use category (since that
requires selecting Public Use and also getting around the dropdown having a
default selection).

This commit does not have any errors for Amenities yet.