Generate and display errors for Amenities
This commit allows the AmenityForm objects to generate and present the errors
that are caused when Amenities fail to save. This is a little more involved
than the comparable commit for Visits, because this involved some work on
Factories as well as some work to change how arguments were passed and how
validations were done (well, named mostly).

Because the `validates` method is aliased to the `valid?` method, we should
also take care in making our classes work correctly. This means aside from
aliasing the `validates` method, we also need to make sure to handle the
`valid?` method. This goes for both VisitForm and AmenityForm.

Also, we originally were going to copy the AmenityForm errors to VisitForm,
but the spec for validations on the VisitForm warned us away from that, as it
basically served only to duplicate errors. And we read the errors off the
AmenityForm when in the view anyway.