Save Request a Visit Form, Page 1
This commit allows the Request a Visit form to save and proceed to page 2
(Team Selection). Turns out that most of the work was done during the
validation step and we just needed to connect the last bits and ensure it
worked with a test.

Right now, because the further pages have not yet been written, the form
proceeds to the homepage instead of the Team Selection page. This is temporary
and will be fixed.