Scope projects to selected type and current user
This commit modifies the projects displayed for selection in the Request a Visit form. When a user selects a `project_type` the one projects that are made available for selection will be those of that type. Additionally, in order for a `project` to be displayed, the `current_user` will need to have an active `team_membership` on that `project`, and have permission to add visits.

This work also ensures that the `project_type` selection is preserved when the form is submitted and results in an error state.