Turn off turbo for Devise forms and links
This commit sets `data-turbo: false` on Devise links and forms as there is not yet an out-of-the-box solution for integrating the two tools. There are some [work arounds](https://gorails.com/episodes/devise-hotwire-turbo) but given that Hotwire will be the default in Rails 7, it is likely that a supported solution will surface in the near future.