Update visit request form
This commit adds `status: :unprocessable_entity` to the `VisitsController#new` action in the event that creating a visit fails. This will allow Turbo to display the form error response. It also removes the `remote: true` declaration from the form, because it is not necessary any more and is considered ‘the old way’ of doing things.

Interestingly, after making this change the system test failed unless a sleep was added to a particular method in the flow object. Suggestions welcome for a less hacky way to handle this.