Trigger event when projects are filtered
This commit changes the form that is used to filter tabular project data by status to `requestSubmit()`, which fires an event and is recognized by Turbo, instead of using `submit()` which just submits the form. 

Interestingly, after making this change the system test failed unless a sleep was added to a particular method in the flow object. Suggestions welcome for a less hacky way to handle this.