Scope reserves to selected project type
This commit modifies the `reserves` displayed for selection in the Request a Visit form. When a user selects a `project_type` the `reserves` that are made available for selection will have a boolean flag that indicates that the selected `project_type` is allowable. Additionally, it also ensures that the selected `reserve` is preserved should the form be submitted unsuccessfully.