Remove turbo tags from layout
This was previously thought to be a requirement in order for Turbo to work. However, that appears to be untrue and the inclusion of the tags generates errors in the console stating that the module is already included. Turbo is confirmed as working without it now, but we are not sure why it was required initially.