Dynamically set amenity usage dates
This commit dynamically assigns the amenities’ `arrives_on` and `departs_on` dates based on the selected visit `start_date` and `end_date`, respectively. In the vast majority of cases, when a user creates a visit, the amenities selected will be for the same dates. So, in order to streamline the process of creating a visit request, the amenities usage dates are set by default.

This is done by creating two new copy controllers and casting them as the existing `CopyController` so the behavior can be used in a situation where two different elements need to be copied from / into on the same page.