Remove errant column reference from migration
The initial migration for this database was based on the schema.rb from RAMS2.
While this was fine as structure, we were using an SQL dump of a database to
populate sample data. This dump contained a column, `ApplicationSubType`, that
was not in the RAMS2 database. This meant it was in the database we were using
for development (and subsequently in our own schema.rb because of how Rails
generates that). When the table it contained was converted to use Rails'
conventions, it was put into the migration with all the others. The problem
was that since it wasn't in the RAMS2 database, it caused an error when
running the migrations "from zero".

This commit removes the column from the migration. While modifying old
migrations is typically disallowed, this is necessary to get it working in all
cases.