Add Start/End times to Amenities on Visit Form
Whe the user selects an Amenity on the Visit Form, they should be able to
select a date and time for the reservation to start and end. Since they can
already select a date, this commit allows them to select a time. The time can
be chosen in 1-hour increments.