Add study area field to visit form
This commit adds the study area field to the visit form's markup and modifies the `@form_values` passed in the `ReserveInputsController` to include a value for `study_area`. The requirements describe this field as being adjacent to, or inline with the 'Special Needs' field, so it will only be visible once a `reserve` has been selected.