Group amenities by group number and label
This commit changes how amenities are displayed on the first page of the Schedule a Visit form. The amenities will now be grouped by the `group_number` and displayed in separate groups as labeled by the `amenity_group_label` that corresponds to that `group_number`.

This work also updates the seeds and specs to support this change.