Display Project Select Options for Project Type
The way we display and select projects for each project type is changing. We
currently have a `public_use_category` when the user selects a Public project.
This commit makes all the project type selections use the same basic mechanism
for selecting a project, and the `ValueProjection` controller changes the
display of the dropdown's label and whether there is a link to create a new
project (hidden for Public).

This commit also, as a result of these changes, removes the public use
category from the form.