Display public projects when Public Use selected
This commit modifies the query to display projects based on `project_type` such that if the `project_type` is `public_use`, we display all public projects and do not scope them via team membership/user permissions. The other `project_type`s do maintain this behavior.

This work also removes the code that previously assigned `project_id` to 0 if the `project_type` was `public_use` because we don't have a default project anymore. The project will be one of the 7 projects that the user selects.