Add markup for the project information modal
This commit adds basic markup for the project new page so that the modal can be rendered, this content is currently limited since the rest will be added across other tickets. It also adds a partial for the modal, and renders it on the page. This work does not include styling, that is coming in a later commit.