Add more data to base seeds
This commit introduces additional data that can be used to populate production-like databases. Previously, the base seeds file only contained a small amount of data that would not allow users to properly test the functionality of the schedule a visit page. These new additions should provide enough information to test all of the features.