Extract steps presentation to StepsPresenter
The StepsPresenter is a presenter that will be composed inside other
presenters since it handles the presentation of the "step" section of various
wizard-style forms. This presentation also involves a partial that renders an
SVG parametrically. This SVG has 3 main states:

* Upcoming, which is an unhighlighted, unfilled circle containing a number.
* Active, which is a highlighted, filled circle, with the number "cut out"
* Completed, which is an unhighlighted, filled circle, with a checkmark "cut out".

Strictly speaking, an unlighlighted, unfilled circle containing a checkmark is
also possible, but we just aren't using it.