Create the changeable skeleton of Projects Form
This commit starts the skeleton of the New Project Form such that each project
type that is selectable at the top of the form can meaningfully present a
significant change in collected data. The forms in question are stored as
`app/views/projects/_{project_type}_form.html.erb` and are loaded into the
main form as Turbo section replacements.