Add markup for discipline section in project form
This commit adds the markup that is required to select a discipline for a project. The disciplines are hardcoded on the Project model and are each represented by a radio button in the view. There is also a text field to collect a plain text entry when the 'Other' discipline is selected.

The validations for the discipline and discipline_other fields were also added as part of this work.