Add markup for project date fields
This commit adds the markup for the project `start_date` and `end_date` fields, which are date fields that leverage the native HTML5 date picker behavior. This allows users to click on a calendar icon within an input and select a date from a calendar widget. It also ensures that the minimum date that can be selected for both fields is today.

This work also permits the date fields in the controller, as well as some others that were previously excluded, for the sake of completeness.