Validate the presence of methods_chemical_list
This commit validates the presence of the `methods_chemicals_list` attribute on the Project model. This will only be required if the `method_chemicals` attribute is truthy. It's worth mentioning that all of the `method_` attributes require the presence of a true/false value but testing for presence is not useful in this case because a value of false would cause a presence validation to fail.