Add markup/behavior to planning section of form
This commit adds the necessary markup to display the planning questions section of the project form, including some Javascript to toggle a class to display a text input if the chemical question is selected. This deviates a bit from performing this via CSS but that proved to be challenging since the input was not (and did not make sense to be) a direct sibling of the 'Yes' radio button.

One thing to note was that in order to use the pre-existing `toggle` stimulus controller, an additional method was added that would only toggle if the target id value matched that of the selected input.