Extract project form styles into a mixin
This commit extracts the project form styles into a mixin that can be used across the various form partials that represent the different tabs of the new project form. Previously, these were scoped under the section with a class of `research` but since these styles will be needed in other sections, this mixin will allow for that with less duplication.