Add remaining, duplicate sections to class tab
This commit adds the remaining form sections to the class tab of the new project form. Much of the markup is duplicated as the sections are the same. Given that the sections were relatively simple and will likely only be duplicated once, it did not seem necessary to extract at this point. In addition to the markup, this also adds the proper conditional validations to the `Project` model.