comments on reserve questions
This is for the VISITS section as this is a RESERVE BASED issue.
This modification is being done to make reserve questions database compatible with the permits record and allow us to add reserve specific permits as reserve questions.
This will allow us to be more flexible with the way reserve want to collect additional information now that RAMS 3 is project based.  
The changes to the reserve questions page in visits will be minor and the same rules used for Location = project will apply as these questions need only be asked once during an initial visit creation.