Migration File for GRANTS database
A few fields were added.  One is sponsor_list so that this pulldown will have a stored location in the table when updating a record.
The old rams2 database will store the historical sponsor data in sponsor_other field in rams 3.
was_denied and funding_opportunity_number are the other 2 additional fields.