ChangePermitsAndPermitAnswers migration file
This change will make permits for the projects.  
These are permits that apply to ALL projects.  as apposed to permits that are reserve based that have been moved to the reserve questions database.
The permits database will store the permit information
the project_permit_answers table will store the YES or NO answer (1,0) from the user.
There are additional fields in the database to help will filtering.  (involves_****)