Renames Project's project types to match enum
The `project_type_options` on Project were using the "full name" of the
options, but the values that will be used in the database (and therefore the
form) will be the "short" names. This commit changes the name of everything
from `university_class` to `class` and from `meeting_or_conference` to
`meeting`, including styles, ids, partial names, and spec names.