Error messages for New Project Form, Page 1
When you submit the Project form and there are errors, we should make sure
that the errors are displayed correctly and well-styled. This commit ensures
that all errors show properly and that the form is otherwise as it was when it
was submitted, including the "popup" fields on the Other discipline and the
Chemicals question under Planning Questions. All three "tabs" are styled in
this manner.