Make project reserve nullable
This commit modifies the projects table to allow a null value for `reserve_id`. In RAMS 2, this association was important but it is no longer necessary due to changes that we are making to the data model. However, there is still data in the system that we need to accomodate so this field won't be removed permanently at this time. This work also adds an `optional: true` option to the association that is defined on the `Project` model.