Add endpoint to submit page 1 of project form to
This commit adds the `Projects::TeamController` that the user will be directed to when the first page of the project form is submitted. Since this is a multi-step form and the project record will have already been created, this controller will grab the `project_id` from params and use it within `edit` and `update` actions to keep this as RESTful as possible.