Fix error state for dates and method questions
This commit fixes several issues that were discovered pertaining to the error state of the form during the submission process. The fields involved were the `method_` and date fields. For the method fields, when submit was clicked, they would all present as “Yes” in the error state which is not correct. That is because there was no setter to convert “Yes” or “No” values into explicitly `true` or `false`. This was however, happening implicitly as “No” was being type cast to true when the record was validated as the string “No” is truthy.

The setters for `start_date` and `end_date` are actually unnecessary because the native html date picker does not allow you to type in invalid date values and the submitted strings are cast automatically on or around the time the record is validated.