Display project team members on teams page of form
This commit displays a table of project team members on the team page of the project form. Each row will contain a team members name, institution, role, project role, active status, and permissions. It also includes a link to edit. The permissions are not editable inline, but use icons to indicate whether or not the team member has the given permission or not.