Make sure styles look right on team form
On the Project Team form, the CSS class did not get changed to be plural and
so didn't look right. This change fixes that. It also replaces the "check" and
"square" icons to look less like checkboxes (which might be considered
interactive) with an unhighlighted checkmark and a dot (to be a placeholder).