Autocomplete Users in Project Team page
When the user types a part of a name into the search field of the Teams page
of the Project form, it should autocomplete with names matching the query
value. This query is done by building a digram index and matching pairs of
letters from the index and the query string.

The autocomplete is also styled in such a way that it will fill available
space (and also in such a way that will not break the styling of the
autocomplete in the User registration page).