Rename project/teams endpoint to project/team_memberships
This commit renames the project/teams endpoint to more accurately reflect the resource that is being modified. Previously,  seemed fine, because we were modifying a project team in its totality, but what we are _actually_ doing is displaying a list of project team memberships and providing the ability to create and/or edit project team membership(s). Therefore, referring to this endpoint as project/team_memberships is a more accurate depiction of what we are doing, and follows RESTful conventions more closely.