Switch from Fulltext search to Regexp search
The MySQL-based Fulltext search wasn't working in a consistent, predicable,
and most importantly desired manner. So in order to make the results more
useful to users -- and to make texting more reliable -- we're backing out of
that method of search and using regular expressions along the same lines as
the Institution search.